AND-9347 Send (TON): cannot send some of ton's tokens if the amount is longer than Long

+++ src/TheOpenNetwork/Payloads.cpp
@@ -15,7 +15,7 @@ using TW::CommonTON::CellBuilder;
 Cell::Ref jettonTransferPayload(
     const Address& responseAddress,
     const Address& toOwner,
-    uint64_t jettonAmount,
+    uint128_t jettonAmount,
     uint64_t forwardAmount,
     const std::string& comment,
     uint64_t query_id
